pyproject.toml
[tool.poetry]
name = "backlogforgames"
version = "0.1.0"
description = "API BacklogForGames"
authors = ["Niolum <al_belov98@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = {extras = ["standart"], version = "^0.24.0.post1"}
asyncpg = "^0.29.0"
sqlalchemy = {version = "^2.0.23", extras = ["asyncio"]}
alembic = "^1.13.0"
python-dotenv = "^1.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
